+++ lib/models/models.dart
@@ -1,3 +1,4 @@
 export 'room_type.dart';
 export 'branding.dart';
 export 'property_details.dart';
+export 'ai_message.dart';
